+++ src/hint-system/analyzable-override.ts
@@ -40,11 +40,11 @@ export class AnalyzableHintMenu {
             },
             focusGained() {
                 this.nameGui.doStateTransition('DEFAULT')
-                HintSystem.g.activateHint(0, this)
+                HintSystem.g.activateHint(this.entity)
             },
             focusLost() {
                 this.nameGui.doStateTransition('HIDDEN')
-                HintSystem.g.deactivateHint(0)
+                HintSystem.g.deactivateHint(HintSystem.g.focusedHE)
             },
             alignGuiPosition() {
                 this.parent()