+++ src/hint-system/climbable-terrain.ts
@@ -137,12 +137,12 @@ export class ClimbableTerrainHints {
             focusGained() {
                 this.parent()
                 this.nameGui.doStateTransition('DEFAULT')
-                HintSystem.g.activateHint(0, this)
+                HintSystem.g.activateHint(this.entity)
             },
             focusLost() {
                 this.parent()
                 this.nameGui.doStateTransition('HIDDEN')
-                HintSystem.g.deactivateHint(0)
+                HintSystem.g.deactivateHint(HintSystem.g.focusedHE)
             },
             alignGuiPosition() {
                 this.parent()