+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added interactable volume slider
+- Added entity hints volume slider
+- Enemy hint sounds, jump hint sounds and hint sounds are now pitched down when they're behind the player
+
 ### Fixed
 
 - Fix potential crash on jump